server/requirements.txt
astroid==2.5.6
click==8.0.0
Flask==2.0.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
#psycopg2==2.8.6
pylint==2.8.2
python-dotenv==0.17.1
SQLAlchemy==1.4.15
toml==0.10.2
Werkzeug==2.0.1
wrapt==1.12.1
